--- a/package.json
+++ b/package.json
@@ -6,7 +6,8 @@
     "build": "npx lerna run build --stream",
     "build-docs": "node bin/build-markdown.js",
     "clean": "npx lerna clean",
-    "lint": "eslint . --ext .js,.ts; npm run lint:types; npm run prettier",
+    "lint": "npm run lint:types; npm run lint:js; npm run prettier",
+    "lint:js": "eslint . --ext .js,.ts",
     "lint:types": "npm run lint:types --if-present --workspaces",
     "prepare": "husky install",
     "prettier": "prettier --check .",
@@ -31,10 +32,9 @@
   "devDependencies": {
     "@commitlint/cli": "^17.6.6",
     "@commitlint/config-conventional": "^17.6.6",
-    "@readme/eslint-config": "^11.0.1",
+    "@readme/eslint-config": "^12.0.0",
     "alex": "^11.0.0",
     "eslint": "^8.44.0",
-    "eslint-plugin-vitest": "^0.2.8",
     "husky": "^8.0.3",
     "prettier": "^3.0.1"
   },