--- a/packages/api/.eslintrc
+++ b/packages/api/.eslintrc
@@ -1,5 +1,4 @@
 {
-  "extends": ["@readme/eslint-config/docs"],
   "globals": {
     "fetch": true,
     "Response": true
@@ -8,22 +7,7 @@
     "ecmaVersion": 2020
   },
   "rules": {
-    // @todo fix these eventually
-    "@typescript-eslint/no-explicit-any": "off",
-
-    // We use `@fixme` in some spots and it's fine.
-    "jsdoc/check-tag-names": "off",
-
-    // Though we aren't requiring JSDoc blocks to be present, if they are they should be properly
-    // formatted.
-    "jsdoc/require-jsdoc": "off",
-    "jsdoc/tag-lines": "off",
-
-    // TypeScript types are better suited for these cases instead.
-    "jsdoc/require-param": "off",
-    "jsdoc/require-param-type": "off",
-    "jsdoc/require-returns": "off",
-    "jsdoc/require-returns-type": "off"
+    "@typescript-eslint/no-explicit-any": "off" // @todo fix these eventually
   },
   "overrides": [
     {