--- a/packages/api/.npmignore
+++ b/packages/api/.npmignore
@@ -4,5 +4,5 @@ test/
 .eslint*
 .gitignore
 .prettier*
-jest.config.js
+vitest.*
 example.js