--- a/packages/api/.npmignore
+++ b/packages/api/.npmignore
@@ -4,5 +4,5 @@ test/
 .eslint*
 .gitignore
 .prettier*
-vitest.*
+vitest*
 example.js