--- a/packages/api/package.json
+++ b/packages/api/package.json
@@ -13,8 +13,8 @@
     "lint:types": "tsc --noEmit",
     "prebuild": "rm -rf dist/; npm run version",
     "prepack": "npm run build",
-    "test": "jest --coverage $(find test -name '*.test.ts' -not -path '*/smoketest.test.ts')",
-    "test:smoke": "npx jest test/cli/codegen/languages/typescript/smoketest.test.ts",
+    "test": "vitest run --coverage",
+    "test:smoke": "vitest run --config=vitest-smoketest.config.ts ",
     "version": "node -p \"'// This file is automatically updated by the build script.\\nexport const PACKAGE_NAME = \\'' + require('./package.json').name + '\\';\\nexport const PACKAGE_VERSION = \\'' + require('./package.json').version + '\\';'\" > src/packageInfo.ts; git add src/packageInfo.ts"
   },
   "repository": {
@@ -75,7 +75,6 @@
     "@readme/oas-examples": "^5.9.0",
     "@types/caseless": "^0.12.2",
     "@types/find-cache-dir": "^3.2.1",
-    "@types/jest": "^29.5.2",
     "@types/js-yaml": "^4.0.5",
     "@types/lodash.camelcase": "^4.3.7",
     "@types/lodash.deburr": "^4.1.7",
@@ -87,14 +86,13 @@
     "@types/semver": "^7.3.13",
     "@types/ssri": "^7.1.1",
     "@types/validate-npm-package-name": "^4.0.0",
+    "@vitest/coverage-v8": "^0.34.1",
     "fetch-mock": "^9.11.0",
-    "jest": "^29.6.1",
-    "jest-extended": "^4.0.0",
     "oas-normalize": "^8.3.2",
-    "ts-jest": "^29.1.1",
     "type-fest": "^3.5.4",
     "typescript": "^4.9.5",
-    "unique-temp-dir": "^1.0.0"
+    "unique-temp-dir": "^1.0.0",
+    "vitest": "^0.34.1"
   },
   "prettier": "@readme/eslint-config/prettier"
 }