--- a/packages/api/test/.eslintrc
+++ b/packages/api/test/.eslintrc
@@ -1,27 +1,6 @@
 {
-  "extends": ["plugin:vitest/all"],
-  "plugins": ["import", "node"],
+  "extends": "@readme/eslint-config/testing/vitest",
   "rules": {
-    "func-names": "off",
-
-    "import/no-extraneous-dependencies": "off",
-
-    "node/no-extraneous-require": "off",
-
-    // Sniff out tests that have useless `async` declarations. Since there's valid usecases for
-    // having a function be async and not return or await a Promise, we're only running this rule
-    // for tests where such a case isn't likely to be a thing.
-    "require-await": "error",
-
-    "vitest/max-expects": "off",
-    "vitest/no-conditional-in-test": "off",
-    "vitest/no-conditional-tests": "off",
-    "vitest/no-hooks": "off",
-    "vitest/prefer-to-be-falsy": "off",
-    "vitest/prefer-to-be-truthy": "off",
-    "vitest/require-top-level-describe": "off",
-
-    "vitest/no-conditional-expect": "off",
-    "vitest/prefer-lowercase-title": "off"
+    "vitest/no-conditional-expect": "off"
   }
 }