--- a/packages/api/test/core/parseResponse.test.ts
+++ b/packages/api/test/core/parseResponse.test.ts
@@ -1,4 +1,5 @@
 import 'isomorphic-fetch';
+import { describe, beforeEach, it, expect } from 'vitest';
 
 import parseResponse from '../../src/core/parseResponse';