--- a/packages/api/test/global.d.ts
+++ b/packages/api/test/global.d.ts
@@ -1,3 +1,1 @@
-import 'jest-extended';
-
-import './helpers/jest.matchers';
+import './helpers/vitest.matchers';