--- a/packages/api/test/helpers/vitest.matchers.ts
+++ b/packages/api/test/helpers/vitest.matchers.ts
@@ -52,7 +52,6 @@ expect.extend({
   },
 
   toHaveCustomUserAgent(headers: string[]) {
-    // const { printReceived } = this.utils;
     const userAgent = headers['user-agent'];
     const pass = userAgent.match(/^api \(node\)\/(\d+.\d+(.\d+|unit-testing))$/);