--- a/packages/api/test/integration.test.ts
+++ b/packages/api/test/integration.test.ts
@@ -3,6 +3,7 @@ import type { OASDocument } from 'oas/dist/rmoas.types';
 import datauri from 'datauri';
 import fetchMock from 'fetch-mock';
 import uniqueTempDir from 'unique-temp-dir';
+import { describe, beforeAll, beforeEach, afterEach, it, expect } from 'vitest';
 
 import api from '../src';
 import Cache from '../src/cache';