--- a/packages/api/test/server.test.ts
+++ b/packages/api/test/server.test.ts
@@ -1,5 +1,6 @@
 import fetchMock from 'fetch-mock';
 import uniqueTempDir from 'unique-temp-dir';
+import { describe, beforeAll, beforeEach, afterEach, it, expect } from 'vitest';
 
 import api from '../src';
 import Cache from '../src/cache';