--- a/packages/httpsnippet-client-api/.npmignore
+++ b/packages/httpsnippet-client-api/.npmignore
@@ -3,4 +3,4 @@ test/
 .eslint*
 .gitignore
 .prettier*
-jest.config.js
+vitest.*