--- a/packages/httpsnippet-client-api/package.json
+++ b/packages/httpsnippet-client-api/package.json
@@ -9,7 +9,7 @@
     "lint:types": "tsc --noEmit",
     "prebuild": "rm -rf dist/",
     "prepack": "npm run build",
-    "test": "jest --coverage"
+    "test": "vitest run --coverage"
   },
   "repository": {
     "type": "git",
@@ -36,14 +36,13 @@
     "@readme/oas-examples": "^5.9.0",
     "@readme/openapi-parser": "^2.4.0",
     "@types/content-type": "^1.1.5",
-    "@types/jest": "^29.5.2",
     "@types/stringify-object": "^4.0.2",
+    "@vitest/coverage-v8": "^0.34.1",
     "api": "file:../api",
     "fetch-mock": "^9.11.0",
     "isomorphic-fetch": "^3.0.0",
-    "jest": "^29.6.1",
-    "ts-jest": "^29.1.1",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "vitest": "^0.34.1"
   },
   "prettier": "@readme/eslint-config/prettier"
 }