--- a/packages/httpsnippet-client-api/package.json
+++ b/packages/httpsnippet-client-api/package.json
@@ -9,7 +9,7 @@
     "lint:types": "tsc --noEmit",
     "prebuild": "rm -rf dist/",
     "prepack": "npm run build",
-    "test": "vitest run --coverage"
+    "test": "vitest --coverage"
   },
   "repository": {
     "type": "git",