--- a/packages/httpsnippet-client-api/test/.eslintrc
+++ b/packages/httpsnippet-client-api/test/.eslintrc
@@ -1,23 +1,3 @@
 {
-  "extends": ["plugin:vitest/all"],
-  "plugins": ["import", "node"],
-  "rules": {
-    "func-names": "off",
-
-    "import/no-extraneous-dependencies": "off",
-
-    "node/no-extraneous-require": "off",
-
-    // Sniff out tests that have useless `async` declarations. Since there's valid usecases for
-    // having a function be async and not return or await a Promise, we're only running this rule
-    // for tests where such a case isn't likely to be a thing.
-    "require-await": "error",
-
-    "vitest/no-conditional-in-test": "off",
-    "vitest/no-conditional-tests": "off",
-    "vitest/no-hooks": "off",
-    "vitest/prefer-to-be-falsy": "off",
-    "vitest/prefer-to-be-truthy": "off",
-    "vitest/require-top-level-describe": "off"
-  }
+  "extends": "@readme/eslint-config/testing/vitest"
 }