--- a/packages/httpsnippet-client-api/vitest.config.ts
+++ b/packages/httpsnippet-client-api/vitest.config.ts
@@ -1,0 +1,8 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    exclude: ['dist', 'node_modules', '**/__datasets__/**', '**/__fixtures__/**', '**/helpers/**'],
+  },
+});