… packages/api/jest.config.js
… packages/httpsnippet-client-api/jest.config.js
… packages/httpsnippet-client-api/test/index.test.ts