## 🧰 Changes

I've been wildy impressed with [Vitest](https://npm.im/vitest) so this is moving us off Jest and over to that. The performance gains aren't going to break the bank, but I'll happily take a couple seconds off of the `api` suite any day.

#### ◀️ Before

##### api
```
Test Suites: 18 passed, 18 total
Tests:       7 todo, 216 passed, 223 total
Snapshots:   0 total
Time:        29.639 s
Ran all test suites matching /test\/core\/index.test.ts|test\/core\/getJSONSchemaDefaults.test.ts|test\/core\/parseResponse.test.ts|test\/core\/prepareParams.test.ts|test\/core\/prepareAuth.test.ts|test\/auth.test.ts|test\/index.test.ts|test\/cache.test.ts|test\/cli\/codegen\/languages\/typescript\/utils.test.ts|test\/cli\/codegen\/languages\/typescript.test.ts|test\/cli\/storage.test.ts|test\/fetcher.test.ts|test\/cache-tmp.test.ts|test\/dist.test.ts|test\/cache-custom.test.ts|test\/config.test.ts|test\/integration.test.ts|test\/server.test.ts/i.
npx jest --coverage   77.67s user 5.05s system 275% cpu 30.033 total
```

##### httpsnippet-client-api
```
Test Suites: 1 passed, 1 total
Tests:       68 passed, 68 total
Snapshots:   0 total
Time:        4.174 s, estimated 5 s
Ran all test suites.
npx jest  6.69s user 0.40s system 149% cpu 4.739 total
```

#### ⏭️ After
##### api
```
 Test Files  18 passed (18)
      Tests  216 passed | 7 todo (223)
   Start at  18:22:46
   Duration  26.88s (transform 817ms, setup 305ms, collect 4.15s, tests 35.72s, environment 2ms, prepare 1.38s)

npx vitest run  56.27s user 4.04s system 220% cpu 27.316 total
```

##### httpsnippet-client-api
```
 Test Files  1 passed (1)
      Tests  68 passed (68)
   Start at  18:24:35
   Duration  3.32s (transform 184ms, setup 0ms, collect 245ms, tests 2.87s, environment 0ms, prepare 52ms)

npx vitest run  5.34s user 0.50s system 155% cpu 3.750 total
```